Cleanup: move pygame code to src/legacy; add deprecation notices; update README for QML primary UI (removed View>Size).
