+++ src/legacy/axis_config_dialog.py
